+++ bin/wasm-node/CHANGELOG.md
@@ -2,8 +2,15 @@
 
 ## Unreleased
 
+### Added
+
+- Add support for parsing the "fron" (Frontier) consensus log items in headers. The content of these log items is ignored by the client.
+
+## 0.6.5 - 2022-17-03
+
 ### Changed
 
+- Chain specifications with a `codeSubstitutes` field containing a block hash are no longer supported ([#2127](https://github.com/paritytech/smoldot/pull/2127)).
 - Prune list of unverified blocks if it grows too much in order to resist spam attacks ([#2114](https://github.com/paritytech/smoldot/pull/2114)).
 - Log block's parent hash in case of block announce ([#2105](https://github.com/paritytech/smoldot/pull/2105)).
 - Only call `console.error` once in case of a Rust panic ([#2093](https://github.com/paritytech/smoldot/pull/2093)).